% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwinkit-package.R
\docType{package}
\name{darwinkit-package}
\alias{darwinkit}
\alias{darwinkit-package}
\title{darwinkit: R interface to the DarwinKit Darwin Core validator}
\description{
Build Darwin Core / OBIS validation configs from data frames with pipeable verbs and validate them against the DarwinKit engine (the dwkt binary).
}
\author{
\strong{Maintainer}: Hakai Institute \email{info@hakai.org}

}
\keyword{internal}
